# ENH: speed up putmask avoiding % in inner loop