@jaimefrio 2015-01-25
Ok, after fixing the bug in the previous code, this is still faster for my original test case:

```
%timeit np.putmask(a, m, b)
1000000 loops, best of 3: 1.56 us per loop
```

So it is 70% faster for a half full mask, although the worst case of an empty mask will result in almost the same timing as above, so a 33% slowdown for that case.

I will take a look at the tests later today, as they obviously need some extending.