@juliantaylor 2015-02-24
oh this isn't merged yet :)
some tests for the bug before would be good, but I don't mind merging it now either