@charris 2015-02-24
Well, then, I'll just merge it as is. Some tests of this functionality would be nice...